Move Kotlin syntax highlighter from lib/llm to components/KotlinCode
